+++ lazycatapp/src/main/java/xyz/mxue/lazycatapp/service/CategoryService.java
@@ -8,7 +8,7 @@ public interface CategoryService {
 
     List<Category> getAllCategories();
 
-    Category getCategoryById(Integer id);
+    Category getCategoryById(Long id);
 
     long count();
 }
